+++ src/Storage/src/Storage/ReadableInterface.php
@@ -1,17 +1,11 @@
 <?php
 
-/**
- * This file is part of Spiral Framework package.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Storage\Storage;
 
 use JetBrains\PhpStorm\ExpectedValues;
+use Psr\Http\Message\UriInterface;
 use Spiral\Storage\Exception\FileOperationException;
 use Spiral\Storage\Exception\InvalidArgumentException;
 use Spiral\Storage\StorageInterface;
@@ -20,7 +14,6 @@ use Spiral\Storage\Visibility;
 /**
  * @psalm-import-type VisibilityType from Visibility
  *
- * @psalm-import-type IdType from StorageInterface
  * @see StorageInterface
  */
 interface ReadableInterface
@@ -28,68 +21,61 @@ interface ReadableInterface
     /**
      * {@see BucketInterface::getContents()}
      *
-     * @param IdType $id
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function getContents($id): string;
+    public function getContents(string|UriInterface|\Stringable $id): string;
 
     /**
      * {@see BucketInterface::getStream()}
      *
-     * @param IdType $id
      * @return resource
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function getStream($id);
+    public function getStream(string|UriInterface|\Stringable $id);
 
     /**
      * {@see BucketInterface::exists()}
      *
-     * @param IdType $id
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function exists($id): bool;
+    public function exists(string|UriInterface|\Stringable $id): bool;
 
     /**
      * {@see BucketInterface::getLastModified()}
      *
-     * @param IdType $id
      * @return positive-int|0
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function getLastModified($id): int;
+    public function getLastModified(string|UriInterface|\Stringable $id): int;
 
     /**
      * {@see BucketInterface::getSize()}
      *
-     * @param IdType $id
      * @return positive-int|0
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function getSize($id): int;
+    public function getSize(string|UriInterface|\Stringable $id): int;
 
     /**
      *{@see BucketInterface::getMimeType()}
      *
-     * @param IdType $id
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function getMimeType($id): string;
+    public function getMimeType(string|UriInterface|\Stringable $id): string;
 
     /**
      * {@see BucketInterface::getVisibility()}
      *
-     * @param IdType $id
      * @return VisibilityType
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
     #[ExpectedValues(valuesFromClass: Visibility::class)]
-    public function getVisibility($id): string;
+    public function getVisibility(string|UriInterface|\Stringable $id): string;
 }