+++ src/Storage/src/Storage/WritableInterface.php
@@ -1,17 +1,11 @@
 <?php
 
-/**
- * This file is part of Spiral Framework package.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Storage\Storage;
 
 use JetBrains\PhpStorm\ExpectedValues;
+use Psr\Http\Message\UriInterface;
 use Spiral\Storage\Exception\FileOperationException;
 use Spiral\Storage\Exception\InvalidArgumentException;
 use Spiral\Storage\FileInterface;
@@ -22,7 +16,6 @@ use Spiral\Storage\Visibility;
 /**
  * @psalm-import-type VisibilityType from Visibility
  *
- * @psalm-import-type IdType from StorageInterface
  * @see StorageInterface
  */
 interface WritableInterface
@@ -30,32 +23,29 @@ interface WritableInterface
     /**
      * {@see BucketInterface::create()}
      *
-     * @param IdType $id
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function create($id, array $config = []): FileInterface;
+    public function create(string|UriInterface|\Stringable $id, array $config = []): FileInterface;
 
     /**
      * {@see BucketInterface::write()}
      *
-     * @param IdType $id
      * @param string|\Stringable|resource $content
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function write($id, $content, array $config = []): FileInterface;
+    public function write(string|UriInterface|\Stringable $id, mixed $content, array $config = []): FileInterface;
 
     /**
      * {@see BucketInterface::setVisibility()}
      *
-     * @param IdType $id
      * @param VisibilityType $visibility
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
     public function setVisibility(
-        $id,
+        string|UriInterface|\Stringable $id,
         #[ExpectedValues(valuesFromClass: Visibility::class)]
         string $visibility
     ): FileInterface;
@@ -63,29 +53,32 @@ interface WritableInterface
     /**
      * {@see BucketInterface::copy()}
      *
-     * @param IdType $source
-     * @param IdType $destination
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function copy($source, $destination, array $config = []): FileInterface;
+    public function copy(
+        string|UriInterface|\Stringable $source,
+        string|UriInterface|\Stringable $destination,
+        array $config = []
+    ): FileInterface;
 
     /**
      * {@see BucketInterface::move()}
      *
-     * @param IdType $source
-     * @param IdType $destination
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function move($source, $destination, array $config = []): FileInterface;
+    public function move(
+        string|UriInterface|\Stringable $source,
+        string|UriInterface|\Stringable $destination,
+        array $config = []
+    ): FileInterface;
 
     /**
      * {@see BucketInterface::delete()}
      *
-     * @param IdType $id
      * @throws FileOperationException
      * @throws InvalidArgumentException
      */
-    public function delete($id, bool $clean = false): void;
+    public function delete(string|UriInterface|\Stringable $id, bool $clean = false): void;
 }